Keep the settings panel presentational: split out ColorFieldPure
The Appearance tab needs a colour picker, and taking it from
`ColorField.tsx` took the whole editor tree with it — that module also
exports the store-connected field, which reaches `Preview` and through it
`ValProvider`. The panel still worked in the app and in Storybook, so the
only symptom was that every jest test touching `SettingsPanel.tsx` died on
a `require` of ESM several layers down, with a stack that named ten files
and not the cause.

`SettingsPanel` is documented as the presentational half. "Does not READ a
store" and "does not IMPORT one" turn out to be different claims, and only
the second is what makes it renderable on its own — so the pure component
gets a file whose boundary says so, and a test that fails the moment a
store, provider or connected field is imported into the panel again.
